+++ Example/PaymentSheet Example/PaymentSheet Example/ExampleEmbeddedElementCheckoutViewController.swift
@@ -45,7 +45,7 @@ class ExampleEmbeddedElementCheckoutViewController: UIViewController {
     private var intentConfig: PaymentSheet.IntentConfiguration {
         return .init(mode: .payment(amount: Int(computedTotals.total),
                                     currency: "USD",
-                                    setupFutureUsage: subscribeSwitch.isOn ? .offSession : nil)
+                                    setupFutureUsage: subscribeSwitch.isOn ? .offSession : nil, captureMethod: .automaticAsync)
         ) { [weak self] paymentMethod, shouldSavePaymentMethod in
             // Create and confirm an intent on your server and invoke `intentCreationCallback` with the client secret or an error.
             // TODO(https://jira.corp.stripe.com/browse/MOBILESDK-2577) Show client-side confirm, not server-side confirm.