+++ Example/PaymentSheet Example/PaymentSheet Example/PlaygroundController.swift
@@ -434,7 +434,7 @@ import UIKit
             switch settings.mode {
             case .payment:
                 return PaymentSheet.IntentConfiguration(
-                    mode: .payment(amount: settings.amount.rawValue, currency: settings.currency.rawValue, setupFutureUsage: nil, paymentMethodOptions: settings.paymentMethodOptionsSetupFutureUsage.makePaymentMethodOptions()),
+                    mode: .payment(amount: settings.amount.rawValue, currency: settings.currency.rawValue, setupFutureUsage: nil, captureMethod: .automaticAsync, paymentMethodOptions: settings.paymentMethodOptionsSetupFutureUsage.makePaymentMethodOptions()),
                     paymentMethodTypes: paymentMethodTypes,
                     paymentMethodConfigurationId: settings.paymentMethodConfigurationId,
                     confirmationTokenConfirmHandler: confirmationTokenConfirmHandler,
@@ -442,7 +442,7 @@ import UIKit
                 )
             case .paymentWithSetup:
                 return PaymentSheet.IntentConfiguration(
-                    mode: .payment(amount: settings.amount.rawValue, currency: settings.currency.rawValue, setupFutureUsage: .offSession, paymentMethodOptions: settings.paymentMethodOptionsSetupFutureUsage.makePaymentMethodOptions()),
+                    mode: .payment(amount: settings.amount.rawValue, currency: settings.currency.rawValue, setupFutureUsage: .offSession, captureMethod: .automaticAsync, paymentMethodOptions: settings.paymentMethodOptionsSetupFutureUsage.makePaymentMethodOptions()),
                     paymentMethodTypes: paymentMethodTypes,
                     paymentMethodConfigurationId: settings.paymentMethodConfigurationId,
                     confirmationTokenConfirmHandler: confirmationTokenConfirmHandler,
@@ -475,7 +475,7 @@ import UIKit
             switch settings.mode {
             case .payment:
                 return PaymentSheet.IntentConfiguration(
-                    mode: .payment(amount: settings.amount.rawValue, currency: settings.currency.rawValue, setupFutureUsage: nil, paymentMethodOptions: settings.paymentMethodOptionsSetupFutureUsage.makePaymentMethodOptions()),
+                    mode: .payment(amount: settings.amount.rawValue, currency: settings.currency.rawValue, setupFutureUsage: nil, captureMethod: .automaticAsync, paymentMethodOptions: settings.paymentMethodOptionsSetupFutureUsage.makePaymentMethodOptions()),
                     paymentMethodTypes: paymentMethodTypes,
                     paymentMethodConfigurationId: settings.paymentMethodConfigurationId,
                     confirmHandler: confirmHandler,
@@ -483,7 +483,7 @@ import UIKit
                 )
             case .paymentWithSetup:
                 return PaymentSheet.IntentConfiguration(
-                    mode: .payment(amount: settings.amount.rawValue, currency: settings.currency.rawValue, setupFutureUsage: .offSession, paymentMethodOptions: settings.paymentMethodOptionsSetupFutureUsage.makePaymentMethodOptions()),
+                    mode: .payment(amount: settings.amount.rawValue, currency: settings.currency.rawValue, setupFutureUsage: .offSession, captureMethod: .automaticAsync, paymentMethodOptions: settings.paymentMethodOptionsSetupFutureUsage.makePaymentMethodOptions()),
                     paymentMethodTypes: paymentMethodTypes,
                     paymentMethodConfigurationId: settings.paymentMethodConfigurationId,
                     confirmHandler: confirmHandler,