+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/PaymentSheet+DeferredAPITest.swift
@@ -65,7 +65,7 @@ final class PaymentSheet_DeferredAPITest: STPNetworkStubbingTestCase {
             mode: .payment(
                 amount: 100,
                 currency: "USD",
-                paymentMethodOptions: .init(setupFutureUsageValues: [.card: .offSession])
+                captureMethod: .automatic, paymentMethodOptions: .init(setupFutureUsageValues: [.card: .offSession])
             )
         ) { _, shouldSavePaymentMethod in
                 // shouldSavePaymentMethod should be true because the intent configuration says to save the card.
@@ -100,7 +100,7 @@ final class PaymentSheet_DeferredAPITest: STPNetworkStubbingTestCase {
                 amount: 100,
                 currency: "USD",
                 setupFutureUsage: nil,  // This should be overridden by PMO SFU
-                paymentMethodOptions: .init(setupFutureUsageValues: [.card: .offSession])
+                captureMethod: .automatic, paymentMethodOptions: .init(setupFutureUsageValues: [.card: .offSession])
             )
         ) { _, _ in return "" }
 
@@ -110,7 +110,7 @@ final class PaymentSheet_DeferredAPITest: STPNetworkStubbingTestCase {
             mode: .payment(
                 amount: 100,
                 currency: "USD",
-                paymentMethodOptions: .init(setupFutureUsageValues: [.card: .onSession])
+                captureMethod: .automatic, paymentMethodOptions: .init(setupFutureUsageValues: [.card: .onSession])
             )
         ) { _, _ in return "" }
 
@@ -120,7 +120,7 @@ final class PaymentSheet_DeferredAPITest: STPNetworkStubbingTestCase {
             mode: .payment(
                 amount: 100,
                 currency: "USD",
-                paymentMethodOptions: .init(setupFutureUsageValues: [.card: .none])
+                captureMethod: .automatic, paymentMethodOptions: .init(setupFutureUsageValues: [.card: .none])
             )
         ) { _, _ in return "" }
 
@@ -132,7 +132,7 @@ final class PaymentSheet_DeferredAPITest: STPNetworkStubbingTestCase {
                 amount: 100,
                 currency: "USD",
                 setupFutureUsage: .offSession
-            )
+            , captureMethod: .automatic)
         ) { _, _ in return "" }
 
         XCTAssertTrue(PaymentSheet.getShouldSavePaymentMethodValue(for: .card, intentConfiguration: intentConfigWithTopLevelSFUOffSession))
@@ -142,14 +142,14 @@ final class PaymentSheet_DeferredAPITest: STPNetworkStubbingTestCase {
                 amount: 100,
                 currency: "USD",
                 setupFutureUsage: .onSession
-            )
+            , captureMethod: .automatic)
         ) { _, _ in return "" }
 
         XCTAssertTrue(PaymentSheet.getShouldSavePaymentMethodValue(for: .card, intentConfiguration: intentConfigWithTopLevelSFUOnSession))
 
         // If nothing is set, don't set SFU
         let intentConfigWithoutSFU = PaymentSheet.IntentConfiguration(
-            mode: .payment(amount: 100, currency: "USD")
+            mode: .payment(amount: 100, currency: "USD", captureMethod: .automatic)
         ) { _, _ in return "" }
 
         XCTAssertFalse(PaymentSheet.getShouldSavePaymentMethodValue(for: .card, intentConfiguration: intentConfigWithoutSFU))
@@ -169,7 +169,7 @@ final class PaymentSheet_DeferredAPITest: STPNetworkStubbingTestCase {
             mode: .payment(
                 amount: 100,
                 currency: "USD",
-                paymentMethodOptions: .init(setupFutureUsageValues: [.card: .offSession])
+                captureMethod: .automatic, paymentMethodOptions: .init(setupFutureUsageValues: [.card: .offSession])
             )
         ) { _, _ in return "" }
 
@@ -186,7 +186,7 @@ final class PaymentSheet_DeferredAPITest: STPNetworkStubbingTestCase {
                 amount: 100,
                 currency: "USD",
                 setupFutureUsage: .onSession
-            )
+            , captureMethod: .automatic)
         ) { _, _ in return "" }
 
         // ...SFU should be set on the params
@@ -201,7 +201,7 @@ final class PaymentSheet_DeferredAPITest: STPNetworkStubbingTestCase {
                 amount: 100,
                 currency: "USD",
                 setupFutureUsage: .offSession,  // This should be overridden by PMO SFU
-                paymentMethodOptions: .init(setupFutureUsageValues: [.card: .none])
+                captureMethod: .automatic, paymentMethodOptions: .init(setupFutureUsageValues: [.card: .none])
             )
         ) { _, _ in return "" }
 
@@ -238,7 +238,7 @@ final class PaymentSheet_DeferredAPITest: STPNetworkStubbingTestCase {
         )
 
         let intentConfig = PaymentSheet.IntentConfiguration(
-            sharedPaymentTokenSessionWithMode: .payment(amount: 1000, currency: "USD"),
+            sharedPaymentTokenSessionWithMode: .payment(amount: 1000, currency: "USD", captureMethod: .automatic),
             sellerDetails: sellerDetails,
             preparePaymentMethodHandler: preparePaymentMethodHandler
         )
@@ -290,7 +290,7 @@ final class PaymentSheet_DeferredAPITest: STPNetworkStubbingTestCase {
         }
 
         let intentConfig = PaymentSheet.IntentConfiguration(
-            sharedPaymentTokenSessionWithMode: .payment(amount: 1000, currency: "USD"),
+            sharedPaymentTokenSessionWithMode: .payment(amount: 1000, currency: "USD", captureMethod: .automatic),
             sellerDetails: PaymentSheet.IntentConfiguration.SellerDetails(networkId: "test_network", externalId: "test_external", businessName: "Till's Pills"),
             preparePaymentMethodHandler: preparePaymentMethodHandler
         )
@@ -392,7 +392,7 @@ final class PaymentSheet_DeferredAPITest: STPNetworkStubbingTestCase {
         }
 
         let intentConfig = PaymentSheet.IntentConfiguration(
-            mode: .payment(amount: 1000, currency: "USD"),
+            mode: .payment(amount: 1000, currency: "USD", captureMethod: .automatic),
             confirmHandler: confirmHandler
         )
 
@@ -430,7 +430,7 @@ final class PaymentSheet_DeferredAPITest: STPNetworkStubbingTestCase {
         }
 
         var intentConfig = PaymentSheet.IntentConfiguration(
-            sharedPaymentTokenSessionWithMode: .payment(amount: 1000, currency: "USD"),
+            sharedPaymentTokenSessionWithMode: .payment(amount: 1000, currency: "USD", captureMethod: .automatic),
             sellerDetails: PaymentSheet.IntentConfiguration.SellerDetails(networkId: "test_network", externalId: "test_external", businessName: "Till's Pills"),
             preparePaymentMethodHandler: preparePaymentMethodHandler
         )