+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/PaymentSheetCustomPaymentMethodTests.swift
@@ -32,7 +32,7 @@ final class PaymentSheetCustomPaymentMethodTests: XCTestCase {
             return .completed
         })
 
-        let intent = Intent.deferredIntent(intentConfig: .init(mode: .payment(amount: 1010, currency: "USD"), confirmHandler: { _, _ in
+        let intent = Intent.deferredIntent(intentConfig: .init(mode: .payment(amount: 1010, currency: "USD", captureMethod: .automatic), confirmHandler: { _, _ in
             XCTFail("Intent confirm handler shouldn't be called")
             return ""
         }))
@@ -104,7 +104,7 @@ final class PaymentSheetCustomPaymentMethodTests: XCTestCase {
             phone: "+15551234567"
         )
 
-        let intent = Intent.deferredIntent(intentConfig: .init(mode: .payment(amount: 1010, currency: "USD"), confirmHandler: { _, _ in
+        let intent = Intent.deferredIntent(intentConfig: .init(mode: .payment(amount: 1010, currency: "USD", captureMethod: .automatic), confirmHandler: { _, _ in
             XCTFail("Intent confirm handler shouldn't be called")
             return ""
         }))
@@ -149,7 +149,7 @@ final class PaymentSheetCustomPaymentMethodTests: XCTestCase {
                 // The merchant's returned result should be passed back in `PaymentSheet.confirm`
                 return merchantReturnedResult
             })
-            let intent = Intent.deferredIntent(intentConfig: .init(mode: .payment(amount: 1010, currency: "USD"), confirmHandler: { _, _ in
+            let intent = Intent.deferredIntent(intentConfig: .init(mode: .payment(amount: 1010, currency: "USD", captureMethod: .automatic), confirmHandler: { _, _ in
                 XCTFail("Intent confirm handler shouldn't be called")
                 return ""
             }))