+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/PaymentSheetExternalPaymentMethodTests.swift
@@ -28,7 +28,7 @@ final class PaymentSheetExternalPaymentMethodTests: XCTestCase {
             return .completed
         })
 
-        let intent = Intent.deferredIntent(intentConfig: .init(mode: .payment(amount: 1010, currency: "USD"), confirmHandler: { _, _ in
+        let intent = Intent.deferredIntent(intentConfig: .init(mode: .payment(amount: 1010, currency: "USD", captureMethod: .automatic), confirmHandler: { _, _ in
             XCTFail("Intent confirm handler shouldn't be called")
             return ""
         }))
@@ -95,7 +95,7 @@ final class PaymentSheetExternalPaymentMethodTests: XCTestCase {
             phone: "+15551234567"
         )
 
-        let intent = Intent.deferredIntent(intentConfig: .init(mode: .payment(amount: 1010, currency: "USD"), confirmHandler: { _, _ in
+        let intent = Intent.deferredIntent(intentConfig: .init(mode: .payment(amount: 1010, currency: "USD", captureMethod: .automatic), confirmHandler: { _, _ in
             XCTFail("Intent confirm handler shouldn't be called")
             return ""
         }))
@@ -138,7 +138,7 @@ final class PaymentSheetExternalPaymentMethodTests: XCTestCase {
                 // The merchant's returned result should be passed back in `PaymentSheet.confirm`
                 return merchantReturnedResult
             })
-            let intent = Intent.deferredIntent(intentConfig: .init(mode: .payment(amount: 1010, currency: "USD"), confirmHandler: { _, _ in
+            let intent = Intent.deferredIntent(intentConfig: .init(mode: .payment(amount: 1010, currency: "USD", captureMethod: .automatic), confirmHandler: { _, _ in
                 XCTFail("Intent confirm handler shouldn't be called")
                 return ""
             }))