+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/PaymentSheetSnapshotTests.swift
@@ -493,7 +493,7 @@ class PaymentSheetSnapshotTests: STPSnapshotTestCase {
     func testPaymentSheetCVCRecollection() {
         stubReturningCustomerResponse()
 
-        let intentConfig = PaymentSheet.IntentConfiguration(mode: .payment(amount: 1000, currency: "USD", setupFutureUsage: .offSession),
+        let intentConfig = PaymentSheet.IntentConfiguration(mode: .payment(amount: 1000, currency: "USD", setupFutureUsage: .offSession, captureMethod: .automatic),
                                                             confirmHandler: confirmHandler,
                                                             requireCVCRecollection: true)
 
@@ -723,7 +723,7 @@ class PaymentSheetSnapshotTests: STPSnapshotTestCase {
     func testPaymentSheet_deferredIntent() {
         stubNewCustomerResponse()
 
-        let intentConfig = PaymentSheet.IntentConfiguration(mode: .payment(amount: 1000, currency: "USD", setupFutureUsage: .offSession),
+        let intentConfig = PaymentSheet.IntentConfiguration(mode: .payment(amount: 1000, currency: "USD", setupFutureUsage: .offSession, captureMethod: .automatic),
                                                             confirmHandler: confirmHandler)
 
         preparePaymentSheet(intentConfig: intentConfig)
@@ -734,7 +734,7 @@ class PaymentSheetSnapshotTests: STPSnapshotTestCase {
     func testPaymentSheet_disableTerms_deferredIntent() {
         stubNewCustomerResponse()
 
-        let intentConfig = PaymentSheet.IntentConfiguration(mode: .payment(amount: 1000, currency: "USD", setupFutureUsage: .offSession),
+        let intentConfig = PaymentSheet.IntentConfiguration(mode: .payment(amount: 1000, currency: "USD", setupFutureUsage: .offSession, captureMethod: .automatic),
                                                             confirmHandler: confirmHandler)
         self.configuration.termsDisplay = [.card: .never]
         preparePaymentSheet(intentConfig: intentConfig)
@@ -748,7 +748,7 @@ class PaymentSheetSnapshotTests: STPSnapshotTestCase {
         stubCustomers()
         stubConsumerSession()
 
-        let intentConfig = PaymentSheet.IntentConfiguration(mode: .payment(amount: 1000, currency: "USD", setupFutureUsage: .onSession),
+        let intentConfig = PaymentSheet.IntentConfiguration(mode: .payment(amount: 1000, currency: "USD", setupFutureUsage: .onSession, captureMethod: .automatic),
                                                             confirmHandler: confirmHandler)
 
         preparePaymentSheet(